pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiomultiprocess==0.9.0",
    "aiosqlite==0.19.0",
    "apscheduler==3.10.4",
    "asyncpg==0.29.0",
    "cryptography==42.0.5",
    "discord-py==2.3.2",
    "emoji==2.10.0",
    "loguru>=0.7.3",
    "python-dotenv==1.0.1",
    "python-telegram-bot==20.6",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "sqlalchemy==2.0.29",
    "telethon==1.32.0",
]
